Facebook coments and dract

+++ src/posts/views.py
@@ -1,18 +1,23 @@
+from urllib import quote_plus
 from django.contrib import messages
 from django.shortcuts import render, get_object_or_404, redirect
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from .models import Post
 # Create your views here.
 from .forms import PostForm
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
 def post_create(request):
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	context = {}
 	if request.method == 'POST':
-		post_form = PostForm(data=request.POST or None)
+		post_form = PostForm(data=request.POST or None, files=request.FILES or None)
 		context["post_form"] = post_form
 
 		if post_form.is_valid():
 			instance = post_form.save(commit=False)
+			instance.user = request.user
 			instance.save()
 			# Message
 			messages.success(request, "successfully Created")
@@ -26,26 +31,45 @@ def post_create(request):
 	context["btn_text"] = "Create"
 	return render(request, "post_form.html", context=context)
 
-def post_detail(request, id=None):
+def post_detail(request, slug=None):
 	context = {}
-	post_instance = get_object_or_404(Post, id=id)
+	post_instance = get_object_or_404(Post, slug=slug)
 	context['title'] = 'Detail'
 	context['post_instance'] = post_instance
+	context['share_string'] = quote_plus(post_instance.content)
 	return render(request, 'post_detail.html', context=context)
 
 def post_list(request):
 	context = {}
-	context['post_list'] = Post.objects.all()
+	post_list = Post.objects.all()
+	
+	paginator = Paginator(post_list, 10) # Show 25 contacts per page
+	page_request = 'page'
+
+	page = request.GET.get(page_request)
+	try:
+		posts = paginator.page(page)
+	except PageNotAnInteger:
+		# If page is not an integer, deliver first page.
+		posts = paginator.page(1)
+	except EmptyPage:
+		# If page is out of range (e.g. 9999), deliver last page of results.
+		posts = paginator.page(paginator.num_pages)
+
 	context['title'] = "List"
+	context['posts'] = posts
+	context['page_request'] = page_request
 	return render(request, 'post_list.html', context=context)
 
-def post_update(request, id=None):
+def post_update(request, slug=None):
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	context = {}
-	post_instance = get_object_or_404(Post, id=id)
+	post_instance = get_object_or_404(Post, slug=slug)
 	post_form = None
 
 	if request.method == 'POST':
-		post_form = PostForm(data=request.POST or None, instance=post_instance)
+		post_form = PostForm(data=request.POST or None, instance=post_instance, files=request.FILES or None)
 		if post_form.is_valid():
 			instance = post_form.save(commit=False)
 			instance.save()
@@ -63,8 +87,10 @@ def post_update(request, id=None):
 
 	return render(request, "post_form.html", context=context)
 
-def post_delete(request, id=None):
-	instance = get_object_or_404(Post, id=id)
+def post_delete(request, slug=None):
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
+	instance = get_object_or_404(Post, slug=slug)
 	instance.delete()
 	messages.success(request, "successfully Deleted")
 	return redirect("posts:list")
